% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktrans.R
\name{KtransModel}
\alias{KtransModel}
\title{Constructs a ktrans model object}
\usage{
KtransModel(defaultParam, parentEnv = .GlobalEnv)
}
\arguments{
\item{parentEnv}{the parent environment for the model}
}
\value{
the model object
}
\description{
Constructs a ktrans model object
}
\examples{
model <- ktrans()
}

