% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KtransParam.R
\name{KtransParam}
\alias{KtransParam}
\title{Constructs a ktrans parameter set object}
\usage{
KtransParam()
}
\value{
the parameter set object
}
\description{
Constructs a ktrans parameter set object
}
\examples{
model <- KtransParam()
}

