% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KtransParam.R
\name{readXML.KtransParam}
\alias{readXML.KtransParam}
\title{Reads a paremeter set from an XML file}
\usage{
\method{readXML}{KtransParam}(ktransParam, file)
}
\value{
nothing
}
\description{
Reads a paremeter set from an XML file
}

