% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KtransParam.R
\name{readXML}
\alias{readXML}
\title{Interface, reads a parameter set from an XML file}
\usage{
readXML(ktransParam, file)
}
\value{
the appropriate function for the class
}
\description{
Interface, reads a parameter set from an XML file
}

